KDatePicker: override changeEvent() instead of setEnabled()
As the comment noted, setEnabled() is not virtual.  Also, there was no
need to separately enabled/disable the individual components, when they
are children of the KDatePicker widget.  Instead, we override
changeEvent() and use it to focus the table (as before).

This also extends the kdatepicktest to temporarily disable the picker.

REVIEW: 117277