KPasswordDialog: let the user change the password visibility
This feature allows the user to change the visibility of the password being typed.
By clicking on the new action appended to the password's QLineEdit, the echo mode will switch between Normal and Password. The two (Breeze) icons used for the action (one for each state) are also included through a .qrc file, so that this feature will be available even for non-Plasma systems where Breeze is not installed.

If a password is given as preset to the dialog, this feature will not be enabled, in order to prevent possible leaks.

REVIEW: 124698
BUG: 224686