Refactor KNewPasswordDialog with a KNewPasswordWidget
A KNewPasswordWidget is now embedded in the dialog, in order to enable the new
features provided by the widget and to remove the duplicated code that has been moved to the new class.

The only new feature that is currently not available from KNewPasswodDialog is the ability to hide
the strength meter bar.

REVIEW: 125619