Refactor and fix the feature introduced in bug 171343
1. Enable this feature only if half step is enabled. Otherwise user can
only select full star, there is no point enable this feature.
2. Fix and merge the logic, especially for the part in mouseMoveEvent.
Previously mouseMoveEvent is calculating the new rating based on current
hover rating, which should use the current rating instead. And the logic
is hard for people to interpret without comment. Thus, this part of code
is refactored to be simpler and more readable with comment.

REVIEW: 126222
CCBUG: 171343