KNewPasswordWidget: update test app and apidox
The code examples are now provided by the @snippet doxygen command, in order to
make sure they always compile and stay updated.

This adds the 'public_example_dir' key in the yaml file, so the @snippet
command will look for files in the tests/ folder. If necessary, the value of
this key can be expanded to a list of folders (see commit 3200bc2aa7 in
kapidox).

REVIEW: 128254