Introduce KToolTipWidget
This new widget is based on the KToolTip code that is currently
duplicated across multiple products: at least Dolphin, systemsettings,
kinfocenter and ktp-contact-list.

Thanks to this new widget it will be possible to apply features/fixes only once.

A new feature that the old code doesn't have is the delayed hide: this
makes it possible to actually use the widget shown in the tooltip.

REVIEW: 129648
CHANGELOG: Add KToolTipWidget, a tooltip that contains another widget.