Fix focus override in KCollapsibleGroupBox
Commit 1076908c34 doesn't work if the widget is a "direct" child of the
groupbox (i.e. the groupbox was passed as parent in the widget contructor).
This happens because the `ChildAdded` event is delivered when the base class
QWidget is created (and QWidget::focusPolicy() is NoFocus by default), but the
child class constructor could set a different focus policy later.

If we don't delay the store+override logic, the widget would always get a NoFocus
policy.

Reviewers: rthomsen, davidedmundson

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D3907