KToolTipWidget: don't take ownership of the content widget
While the layout of the tooltip sets the tooltip as parent
of the content widget, we don't really need to take ownership of it.
We can just restore the old parent once we are done.

This prevents a double delete crash if someone (by accident) deletes
first the tooltip and then the content.

REVIEW: 129849