KSqueezedTextLabel: Respect indent, margin and frame width
Summary:
Text could display cut off when setting indent, margin and/or frame of
the label. On top of that, even when following the size hint to display
the complete text, text could be elided.

This is fixed by taking all chrome into account when determining elision
width and when returning the size hint.

Depends on D7163

Test Plan: Previously expect-failing autotests now pass.

Reviewers: #frameworks, cfeck, dhaumann

Reviewed By: dhaumann

Subscribers: cfeck, dhaumann

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D7164