KAcceleratorManager: set icon text on actions to remove CJK markers
Summary:
This replaces the KToolBar event filter hack to solve the same issue:
when an action text appears in a menu we want the & accelerator, while in
toolbars wewant that removed. Qt takes care of it, except for the more tricky
case of CJK markers: "<chinese here> (&O)" where &O exists only to get an ascii
accelerator.
Instead of hacking the text at painting time (!) it's much more robust to
remove " (&O)" from action texts and sett hat as the icon text upfront.

With this in, we can remove the KToolBar hack which leads to endless repaints.
BUG: 365050
BUG: 377859

Test Plan: Unittest

Reviewers: mardelle, ilic, sandsmark

Subscribers: #frameworks

Differential Revision: https://phabricator.kde.org/D7964