KSqueezedTextLabel: Squeeze text when changing indent or margin
Summary:
When the label is not able to resize in a given layout and calls to
`setIndent` or `setMargin` further reduce the space available, the text
should be squeezed. Qt does not always send a resize event which would notify
us to squeeze the text. Therefore, these functions need
reimplementations adding manual calls to `squeezeTextToLabel`.

This fixes the autotest `testChrome` failing only on the CI (see T6982).

Test Plan:
- `ksqueezedtextlabelautotest` passes locally and with `xvfb-run -s '-screen 0 800x600x24'`
- `abi-compliance-checker` suggests the changes are binary compatible

Reviewers: #frameworks, dhaumann, cfeck

Reviewed By: cfeck

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D7977