Avoid inconsistent passworddialog
Summary:
In case the user starts the KPasswordDialog without the KPasswordDialog::ShowUsernameLine,
the dialog will look a bit odd as the text will still state;
 "Supply a username and password below."
while there is no username field to fill in.

This patch adds a string and sets it in the constructor on the "prompt" to match the flags passed into the constructor.

Test Plan:
{F5499508}

{F5499509}

Reviewers: dfaure, cfeck, ngraham

Reviewed By: dfaure, ngraham

Subscribers: ngraham, #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D8791