KDateComboBox: fix dateChanged() not emitted after typing a date
Summary:
When editing an event in korganizer, if I was typing digits
to edit a date, rather than using another method (date popup; key_up;
etc.), then the change wouldn't be saved.

I would say that on focus-out or Key_Enter, we want to "commit"
the edited date into the final date as desired by the user, by emitting
dateEntered() and dateChanged(), to match the other edit methods.

The documentation clarifies what I think was the intent of the 3
signals, and the unittest checks that the signals are emitted
accordingly.

BUG: 364200

Test Plan: Unittest passes; korganizer use case is fixed.

Reviewers: cfeck, dvratil, mlaurent

Reviewed By: mlaurent

Subscribers: mlaurent, anthonyfieroni, #frameworks, #kde_pim

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D8908