[API dox] Tell doxygen "e.g." does not end the sentence, use ".\ "
Doxygen "autodetects" the end of the brief section by grepping for the
first ".". It needs some help to know which ones to skip.

From http://www.doxygen.org/docblocks.html:
"If you enable this option and want to put a dot in the middle of a
sentence without ending it, you should put a backslash and a space
after it."