add a new property headerVisible
Summary:
add a new property in KPageWidgetItem, headerVisible
this way the page can control whether or not make its own header
visible.
This is necessary for the new systemsettings in which the new
QML kcms display their own title, but the old QWidget based ones
don't and have the KPageWidget display it for them.

Test Plan: tested in systemsettings

Reviewers: #plasma, #frameworks, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: abetts, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D15099