[KMessageBox] Only allow selecting text in the dialog box using the mouse
Summary:
Right now, when you select text in the dialog box, an insertion point appears. It never
disappears. This looks odd: {F7882083}

The insertion point allows you to navigate the text using the keyboard. But this is kind
of pointless since you have to use the mouse to make the insertion point appear in the
text in the first place, so you might as well just use the mouse to make the selection.

This patch makes the text only mouse-selectable. Dolphin recently did the same thing
for its Information Panel labels in D25241 for the same reason.

BUG: 416204
FIXED-IN: 5.67

Test Plan:
Try to open 6 items at once in Dolphin and click-and-drag in the message box text.
Test is still selectable, but an insertion point is no longer created that cannot be
removed.

Reviewers: dfaure, #frameworks, #vdg, ndavis

Reviewed By: #vdg, ndavis

Subscribers: broulik, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D26649