[KMessageWidget] Allow the style to change our palette
Summary:
After we set our palette give the current QStyle a chance to give us a better
palette that maybe fits better because we can only use hardcoded colors. To
avoid infinite recursion when the style calls setPalette on the widget we don't
react to palette change events anymore but when the application palette is
changed.

Test Plan: N/A

Reviewers: #frameworks, apol

Reviewed By: apol

Subscribers: apol, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D27038