[KFontChooser] Remove NoFixedCheckBox DisplayFlag, redundant
Summary:
This partially reverts 0004e5c89a248a508a; the NoFixedCheckBox flag is
redundant as just hiding the "show monospaced fonts only" checkbox when
the FixedFontsOnly flag is set is enough to handle the Konsole use case
(which was the primary reason that flag was introduced in the first place).
Since KF 5.70 isn't released yet, that flag can be safely removed; no
need to clutter the code with redundant flags, and more flags can be
added in the future if there's an actual need/use-case for them.

Thanks to Kurt Hindenburg for catching this,
https://invent.kde.org/kde/konsole/-/merge_requests/84

Test Plan: make && ctest

Reviewers: #frameworks, dfaure, cfeck, bport

Reviewed By: dfaure

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D29219