KTitleWidget: move from QPixmap property to QIcon property
This adds two new properties "icon" and "iconSize" to the class
and deprecates the "pixmap" property.

Some setter methods had already be changed to be about "icon", but
getters & property have been still about a pixmap.

Also rendering the icon into a pixmap at the set call time loses all
the advantages of having QIcon as a on-request pixmap factory, to render
matching the current palette (think Breeze monochrome styled icons) or
adapting to a changed style and its icon sizes.