KFontRequester: remove, the now redundant, nearestExistingFont helper
This method was previously copied/duplicated to the fonts KCM; and has been
removed there for some time to no ill effect that I can see. For the
rationale behind removing it see:
http://commits.kde.org/plasma-desktop/0325d698181055cdaaa93b24ee80172132822d35

This solves an issue with konversation:
- Open settings -> configure konversation -> interface -> fonts, enable
  "show sender nicknames bold in the chat view", OK/Apply
- The <sender> nickname should now become bold in the chat view
- Open the same dialog again, go to the same settings page and enable "chat
  text", OK/Apply; note that the sender nickname has become not-bold, if
  you restart konversation, the nickname is bold again as expected.

This happens because nearestExistingFont gets a QFont that has the
StyleName property set, even for Regular-like StyleName, which means that
QFont::setBold() doesn't work properly.