new setUsernameContextHelp on KPasswordDialog
This adds a new optional UI element next to username field when set.
It's effectively a very user visible whats-this button letting the user
easily discover that there's more information available.

We need this to tell our users how SMB logon information works. It's the
same as Windows' but Linux users tend to not know the various option for
Windows logon names. This new feature allows us to not clutter the UI
too much but still give the user the ability to easily discover input
options that may use.

This is a qwidget-like implementation of a similarly functioning hint
idiom found in the QML based fonts KCM.

The help button is not shown unless a help is set. The help is actually
set on the user lineedit itself as a whatsthis string and then shown as
actual whatsthis window upon button click.