Remove pointless tooltip for yes and no KStandardGuiItems
These buttons have tooltips with the exact same text that is shown on
them, which is pointless.

If this is not a mistake, I suspect that logic is that the application
might override the text in a way that's unclear to the user, so the
tooltip could provide clarity. However this too is false: in such a
situation, the words "yes" and "no" cannot be guaranteed to map to
clearly identifiable actions, so showing these tooltips would only muddy
the situation even further.

Let's trust application developers to write good text if they override
the default text, and remove these tooltips.

BUG: 424658
FIXED-IN: 5.81