Fix crash in KSelectAction
If the destroyed() signal from QComboBox, which is connected to the
comboBoxDeleted() slot, is emitted while the KSelectAction object is being
taken down, there is a crash is QList::removeAll(). Fix the issue by
disconnecting the signal in the KSelectActionPrivate() destructor.

This fixes a crash when closing a window that contains some KSelectAction
objects (e.g. KMail composer window, and OKular).

Add a comment explaining why the KSelectActionPrivate() destructor is virtual,
the comment is taken from an old commit by pino in kdelibs circa 2007 :)

Add unit test.

BUG: 436808