KPasswordLineEdit: honor revealPasswordAvailable also for existing passwords
The KDE HIG recommends:
"
Provide a “Show password” checkbox to unmask the password both when
setting new and when entering existing passwords.
"

Also does the API dox of KPasswordLineEdit not specify that feature
being exclusive when starting with an empty password, it just states
"Whether to show the visibility trailing action in the line edit."

Enabling the feature for existing passwords also yields consistency
with the logic of Plasma/Kirigami QtQuick components.