KFontChooser: fix selecting style name in the chooser
familyListWidget->setCurrentRow() would select the first style in the list,
then later on the correct style is selected, the issue is that the call to
setCurrentRow() has already caused m_selectedFont to change; by moving the
the styleIdentifier() call up, we get the styleId we want.

Also change the test app to select a font with an unusual style name.

NO_CHANGELOG