Revert "Set default buttons"
This reverts commit 54d85c9898f9f23084b4bd47e262e712385aca30.

With a heavy heart, we have to revert this for now because various apps
were relying on the previous implicit assignment of default buttons, and
setting them here breaks people's muscle memory and introduces incorrect
behavior changes, such as making "Cancel" the default button in a
"Save/Don't Save/Cancel" dialog. We can revisit the situation in KF6
when it becomes okay to make breaking changes because apps are expected
to have to adapt to them.

BUG: 442457