API dox: fix generated brief sentences of class and text() method
To work-around doxygen auto-detecting the end of sentences which it does
to see the brief description, escape the space after "e.g." to not have
that ". " be seen as sentence end.

And give text() a short brief sentence, also to avoid running into the e.g. later
in the description with the same problem.

GIT_SILENT