Correct initial values if we don't have a parent yet
This corrects the initial size for KMessageWidget if we
display the widget together with the mainwindow in one go.

For instancee, if you add an option to dolphin commandline
that enables the message widget you will have the correct
sized message without text or buttons.

picture attached on the merge request.