Add KDatePickerPopup
This is a popup menu wrapping a KDatePicker as well as optionally a number
of common shortcuts ("today", "tomorrow", etc).

This is 18 years old and existed as slightly diverged copies in a number
of places (akonadi-contacts, calendarsupport, zanshin and skrooge). A very
similar version also exists in here as part of KDateCombobox. Let's finally
stop that and have a single shared version here.

As a start this is based on the apparently most evolved variant from
calendarsupport, with the few simple i18n() uses replaced by tr(). Merging
with the code in KDateCombobox will be done on top of that by a few
refactoring steps.