Avoid usage of ambiguous two-digit year in date picker
This makes the date picker show the date in short date format but with
a four-digit year. Moreover, it adds support for entering a date in this
format. This solves the problem that two-digit years are ambiguous and
that Qt is stuck in the last century (probably for good reasons), so that
it interprets the two-digit year 22 as meaning 1922 which is highly
unlikely in the year 2022.