Deprecate KMessage*::sorry
When introduced in 1999, it was meant to be used for "small problems",
whereas "error" messages were meant only for "important things":
"
    If your program detects the action specified by the user is somehow
    not allowed, this should never be reported as with error(). Use sorry()
    instead to explain to the user that is action is not allowed.
"

This concept of a light-weight error type for "Not allowed" has not made
it into other recent user-querying KF API, like KMessageWidget or
KNotification. Other toolkits are also not known to feature a similar thing.

Then there are some principle issues, at least with the current design:

A "warning" sign should be used before a problem is hit, not after when
it is too late.

The "Sorry" text is on the border of (not only) KDE HIG, which asks
"
    Use an informal and friendly tone, but not too informal or humorous.
"
where it might be a tad too informal for some in a UI.