Make Full Screen action use consistent text and icon
This is a checkable action, and right now when unchecked, it shows "Full
Screen Mode" but when checked, it shows "Exit Full Screen Mode". The
icon changes too.

Changing the text when checked does not make sense and violates the HIG
(https://develop.kde.org/hig/components/editing/togglebutton/#behavior).

BUG: 240854
FIXED-IN: 5.101