Fix file rating being unreadable in certain themes
Typically, a rating on a file is represented by five stars in a
row. If the first 3 stars are filled and the other 2 aren't, that
represents a 3/5 star rating.

When a theme only provides the filled stars (i.e. "rating") icon
but no unfilled stars (i.e. "rating-unrated") icon, KRatingPainter
would always paint all five stars as filled no matter if the actual
rating was 1/5 or 5/5 because the automatic fallback of the
"rating-unrated" icon is the "rating" icon as stated by the icon
theme spec.

This commit fixes this by having KRatingPainter use a generated
replacement "rating-unrated" icon (a de-saturated and grey-scaled
"rating" icon), if the theme does not provide a "rating-unrated"
icon already.

BUG: 339863


(cherry picked from commit e5ef8184db0733b944cb7a3c9ee1cf239035921b)