KPageDialog: Add search
This change adds the ability to search inside the page dialog. This can
be really helpful if there are a lot of items. The search doesn't just
narrow down the list but also tries to highlight where the text has
matched thus making it easier to find things

This feature is only available if the KPageDialog's underlying view type
is QListView or QTreeView because these two implement row hiding.

This feature was originally added in Kate by me.