KCapacityBar: Consider empty text invisible
Prior to this commit the Breeze style would leave empty space next
to the capacity bar because the capacity bar was reporting that
there should be visible text. That's not true though when no text
has been set which is the default. With this commit text is only
considered visible when text actually exists, which in turn means
that no empty space will be reserved unnecessarily.