Introduce KContextualHelpButton
KContextualHelpButton allows to hide text of any length behind a
small icon-only button. Hovering over the button with a mouse
cursor or pressing the button will show the text. It can contain
images and links. It is accessible by keyboard and to screen
readers.

This is the QtWidgets equivalent to the popular QML
ContextualHelpButton which I implemented a few years ago. Since
then it has found a lot of use in KDE's QML software, especially on
settings pages. It is even being added to our new Human Interface
Guidelines.

This commit makes sure we also have a very similar component
available in QtWidgets.