KTitleWidget: do not autoFillBackground by default, drop StyledPanel frame
The use of autoFillBackground is inconsistent in behaviour of a plain
QLabel (which could be considered a sibling widget type), that paints
the text or pixmap simply onto whatever background the container
widget uses.
Both Breeze & Oxygen even have special code to overwrite this.
ba2252ecdb7d099ff49f46c91c8bdd3e08f34a45 before tried to work-around
the visual effects in prominent usages by defaulting to another
backgroundRole (Base -> Window) but that just moved issues to other
usages with other container backgrounds.

The hard-coded use of a StyledPanel frame also is not in balance with
the default QLabel behaviour. With both Breeze & Oxygen both drawing
nothing with StyledPanel, other than other styles, a frame around the
title widget might come as surprise there when implementing the UI with
KTitleWidget.
And since d3964629cbff90f243c7fd43e41ecdf919c4af1b changed
the contents margins to 0, this pushes the official frame rect, which
is calculated to be around the content, also outside the QFrame widget,
and thus clipped on rendering. So no frame should be seen anyway.

The remove of the QFrame also effectively solves a glitch with the
Fusion style in scaled rendering (QTBUG-126009) and thus the need to
work-around it:
BUG: 475898
FIXED-IN: 6.4.0