Fix build against a static Qt
For a static build Qt generates an object target with the QRC
initialization code that will be linked into consuming code.
We need to take care of installing this though, as well as adding
it to the CMake export set.

Depends on https://invent.kde.org/frameworks/extra-cmake-modules/-/merge_requests/464.