Fix unit tests for ktimecombobox
Since Qt 6.8, QLocale().timeFormat() returns always Ap so the previous
code would always try to convert the local specific AM/PM specifier
to upposer case e.g. "π.μ." "μ.μ." to "Π.Μ." "Μ.Μ." wich would not match
the example anymore and then it would not be converted to the 'x' mask
character and prevent the user to set a time with "π.μ." and only with
"μ.μ." converting the date from midnight to noon.