kpageview: disable invisible search field
kpageview sets a focusproxy to the search field, which is hidden in case
it does not make sense for a particular FaceType. The problem is that
next/prevInFocusChain() test for visibility (and enabledness) of the
candidate focus item, but only for enabledness of the focusproxy.
Allowing this is presumably reasonable behavior, as there may be cases
where an invisible focusproxy should still handle focus for a visible
proxee. It's not reasonable in this case though; the search field is
disabled because it does nothing, so it should not take keyboard focus.

This change sets enabled on the qlineedit as visibility on the search
field is set.