Introduce KIconWidget
It renders a QIcon.

QLabel can only display a fixed pixmap. In order to facilitate
high-dpi and mixed-dpi screens, this widget takes a QIcon and
renders it using the appropriate devicePixelRatio, automatically
updating as needed. It also updates when the palette changes.

An auto test is added that uses the hint.svg in this repo to
not rely on a proper icon theme being presend.

A manual test is added that lets you verify that changing
system palette works.