KPageView: Use correct icon mode if high-contrast color scheme is in use
If a high-contrast color scheme is in use then we need to use selected
icon mode for selected list view items even if the list view doesn't have
focus. This ensures that the correct color is used for icon recoloring
to avoid bad contrast to the background used for selected items.

This is only done for high-contrast color schemes because Breeze Light
uses a light text color for selected list view items that have focus
and a dark text color for selected list view items that don't have focus.