Disable auto-repeat for "toggle full screen" action
Part of:
- https://invent.kde.org/plasma/plasma-desktop/-/issues/149
- https://invent.kde.org/teams/accessibility/collaboration/-/issues/30

Because auto-repeat is on by default, if you hold down F11, the active
window will start to flash at the display's refresh rate. Especially if
 it's got a different color from other windows, the effect can be
really disorienting. This has the potential to cause seizures in people
sensitive to that kind of thing.

Disable auto-repeat to prevent that; now entering or exiting from
fullscreen each require a discrete press-and-release of the F11 key.