KPageTabbedView: Avoid focus going to intermediary container
When using the Tab key to cycle through the widgets in a KPageView
with faceType() == FaceType::Tabbed, the focus would sometimes
"vanish" without a visibly focused object. The cause for this is
that the KPageTabbedView itself receives focus, even though it is
merely a container widget around the real QTabWidget.

This commit sets the focusProxy() of the KPageTabbedView to be the
QTabWidget, which causes the focus to never end up on to the
KPageTabbedView itself, but on the QTabWidget. Considering that the
QTabWidget had already been in the focus chain previously, and it
won't be added redundantly, this effectively reduces the total
count of objects the keyboard focus can go to by 1.

So, the keyboard focus order is more sane now.