KRecentFilesMenu: do not cut text for zero-width screens
When calculating the maximum available width for file names/paths,
RecentFilesEntry uses the narrowest screen reported by
QApplication.screens(). This is usually fine, but on some systems, that
method can return screens with zero width. (For example, when an
xorg.conf ServerLayout refers to a Screen that is not currently
connected.) Consequently, KRecentFilesMenu populates itself with file
names truncated to zero characters. The menu looks like this:

    []
    []
    []

...when it should look like this:

    stuff.zip [/home/user/stuff.zip]
    my.tar.gz [/home/user/my.tar.gz]
    stuff.tar [/home/user/stuff.tar]

This patch makes titleWithSensibleWidth() ignore zero-width screens when deciding how short to cut file names and paths.

BUG: 509320